# Optimize VisitedList counter reset