```console
$ cargo bench -p segment --bench hnsw_search_graph # old
hnsw-index-search-group/hnsw_search
                        time:   [386.29 µs 388.57 µs 391.12 µs]

$ cargo bench -p segment --bench hnsw_search_graph # new
hnsw-index-search-group/hnsw_search
                        time:   [360.82 µs 363.07 µs 365.57 µs]
                        change: [-7.0677% -5.5386% -3.9408%] (p = 0.00 < 0.05)
                        Performance has improved.     
```